FHF speed up

+++ +pio/picrd.m
@@ -3,6 +3,7 @@ picn=length(inpic);
 I=cell(picn);
 for i=1:length(inpic)
     inpic{i}=['\+PIC\',inpic{i},'.bmp'];
-    I{i}=imread(inpic{i});
+    I{i}=double(imread(inpic{i}));
 end
+
 end
